rename MaterialSelect to MaterialSelectInput
